+++ app/src/main/java/com/kix/assessment/modal_classes/Modal_Country.java
@@ -1,24 +1,24 @@
 package com.kix.assessment.modal_classes;
 
-import com.google.gson.annotations.SerializedName;
-
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
+
+import com.google.gson.annotations.SerializedName;
 
 @Entity(tableName = "Country")
 public class Modal_Country {
 
     @PrimaryKey
     @SerializedName("countryId")
-    private Integer countryId;
+    private long countryId;
     @SerializedName("countryName")
     private String countryName;
 
-    public Integer getCountryId() {
+    public long getCountryId() {
         return countryId;
     }
 
-    public void setCountryId(Integer countryId) {
+    public void setCountryId(long countryId) {
         this.countryId = countryId;
     }
